+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
+## [1.1.0] - 2024-08-27
+
 ### Added
 
 - LivePlot: Add `legend` shorthand function
@@ -25,4 +27,5 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Unit tests for the Sketch class
 
 [unreleased]: https://github.com/stephane-caron/matplotlive/compare/v1.0.0...HEAD
+[1.1.0]: https://github.com/stephane-caron/matplotlive/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/stephane-caron/matplotlive/releases/tag/v1.0.0